Actually, selectors should be floats since the app only sees two decimals.

+++ src/main/java/com/quantumslots/physhacks/controllers/HomeController.java
@@ -24,8 +24,8 @@ public class HomeController {
     public void changePotential(Potentials potential){}
     public void measure(double time){
         double position = potential.makeMeasurement(time);
-        double selector1 = player.getSelector1position();
-        double selector2 = player.getSelector2position();
+        float selector1 = player.getSelector1position();
+        float selector2 = player.getSelector2position();
 
         if (selector1 <= position && position <= selector2) {
             // what are gains
